feat(web)!: drop V1 schedule editor
